requirements.txt
asgiref==3.8.1
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
dnspython==2.7.0
mongoengine==0.29.1
Pygments==2.18.0
pymongo==4.10.1
sqlparse==0.5.2
tzdata==2024.2
